Convert generate-types.rs into multiple modules
